Add empty favorites component
